+++ test/Candoumbe.MiscUtilities.UnitTests/StringSegmentExtensionsTests.cs
@@ -5,66 +5,55 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Linq.Expressions;
-using Bogus;
 using FluentAssertions;
+using FsCheck;
+using FsCheck.Xunit;
 using Microsoft.Extensions.Primitives;
 using Xunit;
 using Xunit.Abstractions;
 using Xunit.Categories;
 
-namespace Utilities.UnitTests
+namespace Candoumbe.MiscUtilities.UnitTests
 {
     [UnitTest]
     public class StringSegmentExtensionsTests(ITestOutputHelper outputHelper)
     {
         private readonly ITestOutputHelper _outputHelper = outputHelper;
 
-        public static IEnumerable<object[]> OccurrencesCases
-        {
-            get
+        public static TheoryData<StringSegment, StringSegment, Expression<Func<IEnumerable<int>, bool>>, string> OccurrencesCases
+            => new()
             {
-                yield return new object[]
                 {
                     new StringSegment("Firstname"),
                     new StringSegment("z"),
-                    (Expression<Func<IEnumerable<int>, bool>>)(occurrences => occurrences != null && !occurrences.Any()),
-                    "There's no occurrence of the search element in the string"
-                };
-
-                yield return new object[]
+                    (occurrences => occurrences != null && !occurrences.Any()), "There's no occurrence of the search element in the string"
+                },
                 {
-                    StringSegment.Empty,
-                    new StringSegment("z"),
-                    (Expression<Func<IEnumerable<int>, bool>>)(occurrences => occurrences != null && !occurrences.Any()),
+                    StringSegment.Empty, new StringSegment("z"),
+                    (occurrences => occurrences != null && !occurrences.Any()),
                     "The source element is empty"
-                };
-
-                yield return new object[]
+                },
                 {
                     new StringSegment("Firstname"),
                     new StringSegment("F"),
-                    (Expression<Func<IEnumerable<int>, bool>>)(occurrences =>
-                        occurrences != null
-                        && occurrences.Exactly(1)
-                        && occurrences.Once(pos  => pos == 0)
+                    (occurrences => occurrences != null
+                                    && occurrences.Exactly(1)
+                                    && occurrences.Once(pos => pos == 0)
                     ),
                     "There is one occurrence of the search element in the string"
-                };
-
-                yield return new object[]
+                },
                 {
                     new StringSegment("zsasz"),
-                    new StringSegment("z"),
-                    (Expression<Func<IEnumerable<int>, bool>>)(occurrences =>
+                    new StringSegment("z"), 
+                    (occurrences =>
                         occurrences != null
                         && occurrences.Exactly(2)
                         && occurrences.Once(pos => pos == 0)
-                        && occurrences.Once(pos  =>pos == 4)
+                        && occurrences.Once(pos => pos == 4)
                     ),
                     "There is 2 occurrences of the search element in the string"
-                };
-            }
-        }
+                }
+            };
 
         [Theory]
         [MemberData(nameof(OccurrencesCases))]
@@ -200,58 +189,127 @@ namespace Utilities.UnitTests
             actual.Should()
                   .Be(expectation, reason);
         }
-
-        [Fact]
-        public void LastOccurence_throws_ArgumentException_is_search_is_empty()
+        
+        [Property]
+        public void Given_source_is_not_null_When_search_is_empty_Then_LastOccurrence_returns_the_length_of_the_source(NonNull<string> sourceGenerator)
         {
             // Arrange
-            Faker faker = new();
+            StringSegment source = sourceGenerator.Item;
 
             // Act
-            Action lastOccurrenceWithSearchEmpty = () => new StringSegment(faker.Lorem.Word()).LastOccurrence(StringSegment.Empty);
+            int lastIndex =  source.LastOccurrence(StringSegment.Empty);
 
             // Assert
-            lastOccurrenceWithSearchEmpty.Should()
-                                         .Throw<ArgumentException>();
+            _ = source.Length switch
+            {
+                > 0 => lastIndex.Should().Be(source.Length - 1),
+                _ => lastIndex.Should().Be(0),
+            };
         }
 
-        public static IEnumerable<object[]> FirstOccurrenceCases
+        [Property]
+        public void Given_source_does_not_have_a_value_When_search_has_a_value_Then_LastOccurrence_should_throw_ArgumentOutOfRangeException(NonNull<string> searchGenerator)
         {
-            get
+            // Arrange
+            StringSegment source = (string) null;
+            StringSegment search = searchGenerator.Item;
+
+            // Act
+            Action callingLastOccurrenceWhenSourceDoesNotHaveAValue = () =>  source.LastOccurrence(search);
+
+            // Assert
+            
+            callingLastOccurrenceWhenSourceDoesNotHaveAValue.Should()
+                .Throw<ArgumentOutOfRangeException>()
+                .WithParameterName("source");
+        }
+
+        
+        [Property]
+        public void Given_source_has_a_value_When_search_does_not_have_a_value_Then_LastOccurrence_should_throw_ArgumentOutOfRangeException(NonNull<string> sourceGenerator)
+        {
+            // Arrange
+            StringSegment source = sourceGenerator.Item;
+            StringSegment search = (string) null;
+
+            // Act
+            Action callingLastOccurrenceWhenSourceDoesNotHaveAValue = () =>  source.LastOccurrence(search);
+
+            // Assert
+            callingLastOccurrenceWhenSourceDoesNotHaveAValue.Should()
+                .Throw<ArgumentOutOfRangeException>()
+                .WithParameterName("search");
+        }
+
+        [Property]
+        public void Given_source_does_not_have_a_value_When_search_has_a_value_Then_FirstOccurrence_should_throw_ArgumentOutOfRangeException(NonNull<string> searchGenerator)
+        {
+            // Arrange
+            StringSegment source = null;
+            StringSegment search = searchGenerator.Item;
+
+            // Act
+            Action callingFirstOccurrenceWhenSourceDoesNotHaveAValue = () =>  source.FirstOccurrence(search);
+
+            // Assert
+            
+            callingFirstOccurrenceWhenSourceDoesNotHaveAValue.Should()
+                .Throw<ArgumentOutOfRangeException>()
+                .WithParameterName("source");
+        }
+
+        
+        [Property]
+        public void Given_source_has_a_value_When_search_does_not_have_a_value_Then_FirstOccurrence_should_throw_ArgumentOutOfRangeException(NonNull<string> sourceGenerator)
+        {
+            // Arrange
+            StringSegment source = sourceGenerator.Item;
+            StringSegment search = (string) null;
+
+            // Act
+            Action callingFirstOccurrenceWhenSourceDoesNotHaveAValue = () =>  source.FirstOccurrence(search);
+
+            // Assert
+            callingFirstOccurrenceWhenSourceDoesNotHaveAValue.Should()
+                .Throw<ArgumentOutOfRangeException>()
+                .WithParameterName("search");
+        }
+
+        
+        public static TheoryData<StringSegment, StringSegment, StringComparison, int> FirstOccurrenceCases
+            => new()
             {
-                yield return new object[]
                 {
                     new StringSegment("Firstname"),
                     new StringSegment("z"),
                     StringComparison.OrdinalIgnoreCase,
                     -1
-                };
-
-                yield return new object[]
+                },
                 {
                     StringSegment.Empty,
                     new StringSegment("z"),
                     StringComparison.OrdinalIgnoreCase,
                     -1
-                };
-
-                yield return new object[]
+                },
+                {
+                    StringSegment.Empty,
+                    StringSegment.Empty,
+                    StringComparison.OrdinalIgnoreCase,
+                    0
+                },
                 {
                     new StringSegment("Firstname"),
                     new StringSegment("F"),
                     StringComparison.OrdinalIgnoreCase,
                     0
-                };
-
-                yield return new object[]
+                },
                 {
                     new StringSegment("zsasz"),
                     new StringSegment("z"),
                     StringComparison.OrdinalIgnoreCase,
                     0
-                };
-            }
-        }
+                }
+            };
 
         [Theory]
         [MemberData(nameof(FirstOccurrenceCases))]