+++ CHANGELOG.md
@@ -6,6 +6,16 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### 🔧 Fixes
+- `StringExtensions.FirstOccurrence(string source, string search)` now returns `0` instead of throwing `ArgumentOutOfRangeException` when `source`
+is not null and  ̀search` is empty.
+- `StringExtensions.LastOccurrence(string source, string search)` now returns `0` when both `source` and  `search` are empty.
+- `StringSegmentExtensions.FirstOccurrence(StringSegment source, StringSegment search)` now returns `0` 
+instead of throwing `ArgumentOutOfRangeException` when both `source` and  `search` are equal to `StringSegment.Empty`
+- `StringSegmentExtensions.LastOccurrence(StringSegment source, StringSegment search)` now returns `0`
+instead of throwing `ArgumentOutOfRangeException` when both `source` and  `search` are equal to `StringSegment.Empty`
+
+
 
 ## [0.13.0] / 2024-07-06
 ### 🚨 Breaking changes